clean qr.tsx -> camera.tsx & FormData.tsx
